move MeasureObservation into Measure.ts; clean up commented code
